front and back end edit added

+++ backend/todo/views.py
@@ -1,5 +1,6 @@
 from django.http import HttpResponse, JsonResponse
 from django.shortcuts import render
+from rest_framework.decorators import api_view
 from django.views.decorators.csrf import csrf_exempt
 from rest_framework.parsers import JSONParser
 from rest_framework.response import Response
@@ -9,8 +10,7 @@ from todo.serializers import TodoSeralizer
 
 # Create your views here.
 
-
-@csrf_exempt
+@api_view(["GET", "POST", "DELETE", "PUT"])
 def to_do(request, pk=None):
     if request.method == "POST":
         data = JSONParser().parse(request)
@@ -25,9 +25,20 @@ def to_do(request, pk=None):
         data = TodoSeralizer(data, many=True)
         return JsonResponse({"data": data.data}, status=200)
 
-    if request.method == "DELETE":
+    if request.method == "DELETE" and pk is not None:
         print("delete")
         data = Todo.objects.get(pk=pk)
         print(data)
         data.delete()
         return HttpResponse(status=204)
+
+    if request.method == "PUT" and pk is not None:
+        print("putting")
+        print("pk: ", pk)
+
+        data_object = Todo.objects.get(pk=pk)
+        serializer = TodoSeralizer(data_object, data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+
+        return JsonResponse({"data": serializer.data}, status=200)
